refactor: extract provider client builder module
