+++ crates/pi-coding-agent/src/main.rs
@@ -8,6 +8,7 @@ mod github_issues;
 mod macro_profile_commands;
 mod observability_loggers;
 mod provider_auth;
+mod provider_client;
 mod provider_credentials;
 mod provider_fallback;
 mod runtime_loop;
@@ -108,6 +109,7 @@ pub(crate) use crate::provider_auth::{
     provider_api_key_candidates_with_inputs, provider_auth_capability, provider_auth_mode_flag,
     resolve_api_key,
 };
+pub(crate) use crate::provider_client::build_provider_client;
 #[cfg(test)]
 pub(crate) use crate::provider_credentials::resolve_store_backed_provider_credential;
 pub(crate) use crate::provider_credentials::{
@@ -2397,103 +2399,6 @@ fn event_to_json(event: &AgentEvent) -> serde_json::Value {
             "is_error": result.is_error,
             "content": result.content,
         }),
-    }
-}
-
-fn build_provider_client(cli: &Cli, provider: Provider) -> Result<Arc<dyn LlmClient>> {
-    let auth_mode = configured_provider_auth_method(cli, provider);
-    let capability = provider_auth_capability(provider, auth_mode);
-    if !capability.supported {
-        bail!(
-            "unsupported auth mode '{}' for provider '{}': {} (set {} api-key)",
-            auth_mode.as_str(),
-            provider.as_str(),
-            capability.reason,
-            provider_auth_mode_flag(provider),
-        );
-    }
-
-    let resolver = CliProviderCredentialResolver { cli };
-    let resolved = resolver.resolve(provider, auth_mode)?;
-    let auth_source = resolved.source.as_deref().unwrap_or("none");
-
-    match provider {
-        Provider::OpenAi => {
-            let api_key = resolved.secret.ok_or_else(|| {
-                anyhow!(
-                    "resolved auth mode '{}' for '{}' did not provide a credential",
-                    resolved.method.as_str(),
-                    provider.as_str()
-                )
-            })?;
-
-            let client = OpenAiClient::new(OpenAiConfig {
-                api_base: cli.api_base.clone(),
-                api_key,
-                organization: None,
-                request_timeout_ms: cli.request_timeout_ms.max(1),
-                max_retries: cli.provider_max_retries,
-                retry_budget_ms: cli.provider_retry_budget_ms,
-                retry_jitter: cli.provider_retry_jitter,
-            })?;
-            tracing::debug!(
-                provider = provider.as_str(),
-                auth_mode = resolved.method.as_str(),
-                auth_source = auth_source,
-                "provider auth resolved"
-            );
-            Ok(Arc::new(client))
-        }
-        Provider::Anthropic => {
-            let api_key = resolved.secret.ok_or_else(|| {
-                anyhow!(
-                    "resolved auth mode '{}' for '{}' did not provide a credential",
-                    resolved.method.as_str(),
-                    provider.as_str()
-                )
-            })?;
-
-            let client = AnthropicClient::new(AnthropicConfig {
-                api_base: cli.anthropic_api_base.clone(),
-                api_key,
-                request_timeout_ms: cli.request_timeout_ms.max(1),
-                max_retries: cli.provider_max_retries,
-                retry_budget_ms: cli.provider_retry_budget_ms,
-                retry_jitter: cli.provider_retry_jitter,
-            })?;
-            tracing::debug!(
-                provider = provider.as_str(),
-                auth_mode = resolved.method.as_str(),
-                auth_source = auth_source,
-                "provider auth resolved"
-            );
-            Ok(Arc::new(client))
-        }
-        Provider::Google => {
-            let api_key = resolved.secret.ok_or_else(|| {
-                anyhow!(
-                    "resolved auth mode '{}' for '{}' did not provide a credential",
-                    resolved.method.as_str(),
-                    provider.as_str()
-                )
-            })?;
-
-            let client = GoogleClient::new(GoogleConfig {
-                api_base: cli.google_api_base.clone(),
-                api_key,
-                request_timeout_ms: cli.request_timeout_ms.max(1),
-                max_retries: cli.provider_max_retries,
-                retry_budget_ms: cli.provider_retry_budget_ms,
-                retry_jitter: cli.provider_retry_jitter,
-            })?;
-            tracing::debug!(
-                provider = provider.as_str(),
-                auth_mode = resolved.method.as_str(),
-                auth_source = auth_source,
-                "provider auth resolved"
-            );
-            Ok(Arc::new(client))
-        }
     }
 }
